pyproject.toml
[project]
name = "personal-knowledge-base"
version = "0.1.0"
description = "Sistema RAG pessoal para notas e documentos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core RAG
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",

    # Ollama
    "ollama>=0.1.6",

    # Ingestao de documentos
    "python-frontmatter>=1.1.0",
    "pdfplumber>=0.10.3",
    "pypdf2>=3.0.1",
    "trafilatura>=1.6.3",
    "beautifulsoup4>=4.12.2",
    "unstructured>=0.12.0",

    # Interface
    "streamlit>=1.31.0",
    "typer>=0.9.0",
    "rich>=13.7.0",

    # Visualizacao
    "plotly>=5.18.0",
    "umap-learn>=0.5.5",
    "scikit-learn>=1.4.0",

    # Utilidades
    "tiktoken>=0.5.2",
    "python-dotenv>=1.0.0",
    "watchdog>=3.0.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.2.0",
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.3",
]

[project.scripts]
pkb = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ============================================================================
# RUFF - Linting e Formatacao
# ============================================================================

[tool.ruff]
target-version = "py310"
line-length = 88
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "chroma_db",
]
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
]
ignore = [
    "E501",   # line too long (formatter cuida disso)
]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
